requirements.txt
google-cloud-bigtable
google-cloud-core
google-cloud-storage
google-cloud-bigquery
pandas
fsspec
gcsfs
pandas-gbq